fix compiler warnings and msvc opt options
mingw-gcc:
```
[132/150] compiling c muon.exe.p/src/platform/windows/run_cmd.c.o
../src/platform/windows/run_cmd.c: In function 'copy_pipes':
../src/platform/windows/run_cmd.c:137:87: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
  137 |                 if (!GetQueuedCompletionStatus(ctx->ioport, &_bytes_read, (PULONG_PTR)&pipe, &_overlapped, 100)) {
      |
```

msvc:
```
cl : Command line warning D9002 : ignoring unknown option '/0d'

../src/toolchains.c(1335): warning C4090: 'function': different 'const' qualifiers

../src/lang/analyze.c(1329): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)

../src/lang/eval.c(564): warning C4245: 'function': conversion from 'int' to 'uint32_t', signed/unsigned mismatch

../src/toolchains.c(1335): warning C4090: 'function': different 'const' qualifiers

D:\Downloads\muon\src\platform\windows\run_cmd.c(158) : warning C4754: Conversion rules for arithmetic operations in the comparison at D:\Downloads\muon\src\platform\windows\run_cmd.c(157) mean that one branch cannot be executed. Cast '(... * 2)' to 'ULONG64' (or similar type of 8 bytes).
```
closes #198

fix fmt issues

+++ src/external/samurai/deps.c
@@ -258,7 +258,7 @@ samu_depsinit(struct samu_ctx *ctx, const char *builddir)
 				samu_warn("invalid size, must be greater than 4: %" PRIu32, sz);
 				goto rewrite;
 			}
-			if (ctx->deps.entrieslen != ~buf[sz / 4 - 1]) {
+			if (ctx->deps.entrieslen != (size_t)~buf[sz / 4 - 1]) {
 				samu_warn("corrupt deps log, bad checksum");
 				goto rewrite;
 			}
